Checking for a query parameter to avoid injecting sim files.
- InAppBrowser: analyze URL to add or not the query parameter to
avoid injecting simulation files.
- Moving utilities functions to utils module: isSameOriginRequest and parseUrl.
